backend/data_collector/requirements.txt
# FastAPI and server
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.7

# HTTP and networking
requests>=2.28.2
aiohttp>=3.8.4
python-dotenv>=1.0.0

# Kafka integration
kafka-python>=2.0.2

# Web scraping
scrapy>=2.8.0
selenium>=4.8.3
beautifulSoup4>=4.12.0

# TOR integration
PySocks>=1.7.1
torpy>=1.1.6

# Utilities
python-dateutil>=2.8.2
tenacity>=8.2.2